backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
requests
langchain
langchain-google-genai
backboard-sdk
elevenlabs
